- 视图和透视矩阵，以及灯光设置都改为uniform buffer中。（灯光每帧设置一次，矩阵改动时才设置） - 灯光上限改为12，太大依旧会报错，暂时不知为何 - 添加Event类 - 添加UniformBuffer类 - 添加DepthFrameBuffer和simpleDepthShader，将用于渲染阴影

+++ src/VertexBuffer.cpp
@@ -1,7 +1,7 @@
 #include "VertexBuffer.h"
 #include "GLCALL.h"
 
-VertexBuffer::VertexBuffer(const void* data, unsigned int size, unsigned int count, bool dynamicDraw) : m_Count(count), m_RendererID(0)
+VertexBuffer::VertexBuffer(const void* data, unsigned long long size, unsigned int count, bool dynamicDraw) : m_Count(count), m_RendererID(0)
 {
 	GLCALL(glGenBuffers(1, &m_RendererID));
 	Bind();
